graduate and student updated

+++ src/main/java/com/example/library/repository/GraduateRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.library.entity.Graduate;
 
-public interface GraduateRepository extends JpaRepository<Graduate,Integer>{
+public interface GraduateRepository extends JpaRepository<Graduate,Long>{
 
 }
